app/requirements.txt
Flask>=3.1.1
azure-identity>=1.23.0
azure-keyvault-secrets>=4.9.0
gunicorn>=23.0.0